+++ packages/patrol/example/android/app/build.gradle
@@ -50,7 +50,6 @@ android {
         targetSdkVersion 33
         versionCode flutterVersionCode.toInteger()
         versionName flutterVersionName
-        multiDexEnabled true
         testInstrumentationRunner "pl.leancode.patrol.PatrolJUnitRunner"
         testInstrumentationRunnerArguments clearPackageData: "true"
     }
@@ -61,24 +60,7 @@ android {
 
     buildTypes {
         release {
-            // TODO: Add your own signing config for the release build.
-            // Signing with the debug keys for now, so `flutter run --release` works.
             signingConfig signingConfigs.debug
-        }
-    }
-
-    flavorDimensions "default"
-
-    productFlavors {
-        dev {
-            dimension "default"
-            resValue "string", "app_name", "DEV Challenge"
-            applicationIdSuffix ".dev"
-        }
-
-        prod {
-            dimension "default"
-            resValue "string", "app_name", "Challenge"
         }
     }
 }
