+++ packages/patrol/example/android/app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
     <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
     <application
-        android:label="@string/app_name"
+        android:label="Patrol example"
         android:name="${applicationName}"
         android:icon="@mipmap/launcher_icon">
         <activity
