+++ packages/patrol_finders/example/android/app/src/main/AndroidManifest.xml
@@ -1,8 +1,5 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          xmlns:tools="http://schemas.android.com/tools" package="pl.leancode.patrol.example">
-
+<manifest xmlns:android="http://schemas.android.com/apk/res/android">
     <uses-permission android:name="android.permission.INTERNET"/>
-
     <application
         android:label="example"
         android:name="${applicationName}"
